+++ variables.tf
@@ -1,12 +1,18 @@
 variable "seller_profile_name" {
   type        = string
-  description = "Unique identifier of the service provider's profile. One of 'seller_profile_name' or 'zside_port_name' is required."
+  description = <<EOF
+  Unique identifier of the service provider's profile. One of 'seller_profile_name' or
+  'zside_port_name' is required.
+  EOF
   default     = ""
 }
 
 variable "name" {
   type        = string
-  description = "Name of the connection resource that will be created. It will be auto-generated if not specified."
+  description = <<EOF
+  Name of the connection resource that will be created. It will be auto-generated if not
+  specified.
+  EOF
   default     = ""
 }
 
@@ -22,8 +28,8 @@ variable "port_name" {
 variable "speed" {
   type        = number
   description = <<EOF
-  Speed/Bandwidth to be allocated to the connection - (MB or GB). If not specified, it will be used the minimum
-  bandwidth available for the specified seller profile.
+  Speed/Bandwidth to be allocated to the connection - (MB or GB). If not specified, it will be used
+  the minimum bandwidth available for the specified seller profile.
   EOF
   default     = 0
 }
@@ -53,8 +59,8 @@ variable "seller_authorization_key" {
 variable "seller_metro_code" {
   type        = string
   description = <<EOF
-  Metro code where the connection will be created. If you do not know the code,'seller_metro_name' can be use
-  instead.
+  Metro code where the connection will be created. If you do not know the code,'seller_metro_name'
+  can be use instead.
   EOF
   default     = ""
 
@@ -69,9 +75,9 @@ variable "seller_metro_code" {
 variable "seller_metro_name" {
   type        = string
   description = <<EOF
-  Only required in the absence of 'seller_metro_code'. Metro name where the connection will be created, i.e.
-  'Frankfurt', 'Silicon Valley', 'Ashburn'. One of 'seller_metro_code', 'seller_metro_name' must be
-  provided.
+  Only required in the absence of 'seller_metro_code'. Metro name where the connection will be
+  created, i.e. 'Frankfurt', 'Silicon Valley', 'Ashburn'. One of 'seller_metro_code',
+  'seller_metro_name' must be provided.
   EOF
   default     = ""
 }
@@ -79,9 +85,10 @@ variable "seller_metro_name" {
 variable "seller_region" {
   type        = string
   description = <<EOF
-  The region in which the seller port resides, i.e. 'eu-west-1'. Required only in cases where you need a specific
-  region of a service provider with several regions per metro. Generally there is only one region per metro, and it
-  will be used the first available region in the metro of the specified seller profile.
+  The region in which the seller port resides, i.e. 'eu-west-1'. Required only in cases where you
+  need a specific region of a service provider with several regions per metro. Generally there is
+  only one region per metro, and it will be used the first available region in the metro of the
+  specified seller profile.
   EOF
   default     = ""
 }
@@ -105,25 +112,28 @@ variable "purcharse_order_number" {
 variable "vlan_stag" {
   type        = number
   description = <<EOF
-  S-Tag/Outer-Tag of the primary connection - a numeric character ranging from 2 - 4094. Required if
-  'port_name' is specified.
+  S-Tag/Outer-Tag of the primary connection - a numeric character ranging from 2 - 4094. Required
+  if 'port_name' is specified.
   EOF
   default     = 0
 }
 
 variable "vlan_ctag" {
   type        = number
-  description = "C-Tag/Inner-Tag of the primary connection - a numeric character ranging from 2 - 4094."
+  description = <<EOF
+  C-Tag/Inner-Tag of the primary connection - a numeric character ranging from
+  2 - 4094.
+  EOF
   default     = 0
 }
 
 variable "zside_port_name" {
   type        = string
   description = <<EOF
-  Name of the buyer's port from which the connection would originate the port on the remote side (z-side). Required
-  when destination is another port instead of a service profile. Usually, if you don't have an existing private service
-  profile, this option offers a simple, streamlined way to set up a connection between your own ports. Not compatible
-  with redundant connections"
+  Name of the buyer's port from which the connection would originate the port on the remote side
+  (z-side). Required when destination is another port instead of a service profile. Usually, if
+  you don't have an existing private service profile, this option offers a simple, streamlined
+  way to set up a connection between your own ports. Not compatible with redundant connections.
   EOF
   default     = ""
 }
@@ -131,8 +141,8 @@ variable "zside_port_name" {
 variable "zside_vlan_stag" {
   type        = number
   description = <<EOF
-  S-Tag/Outer-Tag of the connection on the Z side. Required if 'zside_port_name' is specified. A numeric
-  character ranging from 2 - 4094.
+  S-Tag/Outer-Tag of the connection on the Z side. Required if 'zside_port_name' is specified. A
+  numeric character ranging from 2 - 4094.
   EOF
   default     = 0
 }
@@ -140,8 +150,8 @@ variable "zside_vlan_stag" {
 variable "zside_vlan_ctag" {
   type        = number
   description = <<EOF
-  C-Tag/Inner-Tag of the connection on the Z side. This is only applicable with 'named_tag'. A numeric character
-  ranging from 2 - 4094.
+  C-Tag/Inner-Tag of the connection on the Z side. This is only applicable with 'named_tag'. A
+  numeric character ranging from 2 - 4094.
   EOF
   default     = 0
 }
@@ -149,31 +159,48 @@ variable "zside_vlan_ctag" {
 variable "named_tag" {
   type        = string
   description = <<EOF
-  The type of peering to set up in case when connecting to Azure Express Route. One of 'PRIVATE', 'MICROSOFT'.
+  The type of peering to set up in case when connecting to Azure Express Route. One of 'PRIVATE',
+  'MICROSOFT'.
   EOF
   default     = ""
+}
+
+variable "additional_info" {
+  type        = list(object({
+      name    = string,
+      value   = string
+    })
+  )
+  description = <<EOF
+  Additional parameters required for some service profiles. It should be a list of maps containing
+  'name' and 'value  e.g. `[{ name='asn' value = '65000'}, { name='ip' value = '192.168.0.1'}]`.
+  EOF
+  default     = []
 }
 
 variable "service_token_id" {
   type        = string
   description = <<EOF
-  Unique Equinix Fabric key shared with you by a provider that grants you authorization to use their interconnection
-  asset from which the connection would originate.
+  Unique Equinix Fabric key shared with you by a provider that grants you authorization to use
+  their interconnection asset from which the connection would originate.
   EOF
   default     = ""
 }
 
 variable "network_edge_id" {
   type        = string
-  description = "Unique identifier of the Network Edge virtual device from which the connection would originate."
+  description = <<EOF
+  Unique identifier of the Network Edge virtual device from which the connection would
+  originate.
+  EOF
   default     = ""
 }
 
 variable "network_edge_interface_id" {
   type        = number
   description = <<EOF
-  Applicable with 'network_edge_id', identifier of network interface on a given device, used for a connection. If not
-  specified then first available interface will be selected.
+  Applicable with 'network_edge_id', identifier of network interface on a given device, used for a
+  connection. If not specified then first available interface will be selected.
   EOF
   default     = 0
 }
@@ -181,8 +208,8 @@ variable "network_edge_interface_id" {
 variable "redundancy_type" {
   type        = string
   description = <<EOF
-  Whether to create a single connection or redundant. Fabric secondary variables will take no effect unless value
-  'REDUNDANT' is specified.
+  Whether to create a single connection or redundant. Fabric secondary variables will take no
+  effect unless value 'REDUNDANT' is specified.
   EOF
   default     = "SINGLE"
 
@@ -194,16 +221,19 @@ variable "redundancy_type" {
 
 variable "secondary_name" {
   type        = string
-  description = "Name of the secondary connection that will be created. It will be auto-generated if not specified."
+  description = <<EOF
+  Name of the secondary connection that will be created. It will be auto-generated
+  if not specified.
+  EOF
   default     = ""
 }
 
 variable "secondary_port_name" {
   type        = string
   description = <<EOF
-  Name of the buyer's port from which the secondary connection would originate. If not specified, and
-  'port_name' is specified, and 'redundancy_type' is 'REDUNDANT', then the value of 'port_name' will be
-  used.
+  Name of the buyer's port from which the secondary connection would originate. If not specified,
+  and 'port_name' is specified, and 'redundancy_type' is 'REDUNDANT', then the value of 'port_name'
+  will be used.
   EOF
   default     = ""
 }
@@ -211,8 +241,8 @@ variable "secondary_port_name" {
 variable "secondary_speed" {
   type        = number
   description = <<EOF
-  Speed/Bandwidth to be allocated to the secondary connection - (MB or GB). If not specified then primary connection
-  speed will be used.
+  Speed/Bandwidth to be allocated to the secondary connection - (MB or GB). If not specified then
+  primary connection speed will be used.
   EOF
   default     = 0
 }
@@ -220,8 +250,8 @@ variable "secondary_speed" {
 variable "secondary_speed_unit" {
   type        = string
   description = <<EOF
-  Unit of the speed/bandwidth to be allocated to the secondary connection. If not specified then primary connection
-  speed unit will be used.
+  Unit of the speed/bandwidth to be allocated to the secondary connection. If not specified then
+  primary connection speed unit will be used.
   EOF
   default     = ""
 
@@ -236,23 +266,26 @@ variable "secondary_speed_unit" {
 variable "secondary_vlan_stag" {
   type        = number
   description = <<EOF
-  S-Tag/Outer-Tag of the secondary connection. A numeric character ranging from 2 - 4094. Required if
-  'secondary_port_name' is specified. 
+  S-Tag/Outer-Tag of the secondary connection. A numeric character ranging from 2 - 4094. Required
+  if 'secondary_port_name' is specified. 
   EOF
   default     = 0
 }
 
 variable "secondary_vlan_ctag" {
   type        = number
-  description = "C-Tag/Inner-Tag of the secondary connection - a numeric character ranging from 2 - 4094."
+  description = <<EOF
+  C-Tag/Inner-Tag of the secondary connection - a numeric character ranging from
+  2 - 4094.
+  EOF
   default     = 0
 }
 
 variable "secondary_seller_authorization_key" {
   type = string
   description = <<EOF
-  Text field used to authorize secondary connection on the provider side. Value depends on a provider service profile
-  used for connection.
+  Text field used to authorize secondary connection on the provider side. Value depends on a
+  provider service profile used for connection.
   EOF
   default = ""
 }
@@ -260,8 +293,8 @@ variable "secondary_seller_authorization_key" {
 variable "secondary_seller_metro_code" {
   type        = string
   description = <<EOF
-  Metro code where the secondary connection will be created. If not specified then primary connection metro code will
-  be used.
+  Metro code where the secondary connection will be created. If not specified then primary
+  connection metro code will be used.
   EOF
   default     = ""
 
@@ -276,8 +309,8 @@ variable "secondary_seller_metro_code" {
 variable "secondary_seller_metro_name" {
   type        = string
   description = <<EOF
-  Metro name where the secondary connection will be created, i.e. 'Frankfurt', 'Silicon Valley', 'Ashburn'. If not
-  specified then primary connection metro name will be used.
+  Metro name where the secondary connection will be created, i.e. 'Frankfurt', 'Silicon Valley',
+  'Ashburn'. If not specified then primary connection metro name will be used.
   EOF
   default     = ""
 }
@@ -285,8 +318,8 @@ variable "secondary_seller_metro_name" {
 variable "secondary_seller_region" {
   type        = string
   description = <<EOF
-  The region in which the seller port resides, i.e. 'eu-west-1'. If not specified then primary connection region will
-  be used.
+  The region in which the seller port resides, i.e. 'eu-west-1'. If not specified then primary
+  connection region will be used.
   EOF
   default     = ""
 }
@@ -294,9 +327,9 @@ variable "secondary_seller_region" {
 variable "network_edge_secondary_id" {
   type        = string
   description = <<EOF
-  Unique identifier of the Network Edge virtual device from which the secondary connection would originate. If not
-  specified, and 'network_edge_id' is specified, and 'redundancy_type' is 'REDUNDANT' then primary edge device will
-  be used.
+  Unique identifier of the Network Edge virtual device from which the secondary connection would
+  originate. If not specified, and 'network_edge_id' is specified, and 'redundancy_type' is
+  'REDUNDANT' then primary edge device will be used.
   EOF
   default     = ""
 }
@@ -304,8 +337,9 @@ variable "network_edge_secondary_id" {
 variable "network_edge_secondary_interface_id" {
   type        = number
   description = <<EOF
-  Applicable with 'network_edge_id' or 'network_edge_secondary_id', identifier of network interface on a given device,
-  used for a connection. If not specified then first available interface will be selected.
+  Applicable with 'network_edge_id' or 'network_edge_secondary_id', identifier of network interface
+  on a given device, used for a connection. If not specified then first available interface will be
+  selected.
   EOF
   default     = 0
 }
@@ -313,9 +347,9 @@ variable "network_edge_secondary_interface_id" {
 variable "secondary_service_token_id" {
   type        = string
   description = <<EOF
-  Unique Equinix Fabric key shared with you by a provider that grants you authorization to use their interconnection
-  asset from which the secondary connection would originate. Required if 'service_token_id' is specified, and
-  'redundancy_type' is 'REDUNDANT'.
+  Unique Equinix Fabric key shared with you by a provider that grants you authorization to use
+  their interconnection asset from which the secondary connection would originate. Required if
+  'service_token_id' is specified, and 'redundancy_type' is 'REDUNDANT'.
   EOF
   default     = ""
 }