+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+## 0.2.0 (April 27, 2022)
+
+FEATURES:
+
+- Supported definition of `additional_info` blocks.
+
 ## 0.1.1 (April 18, 2022)
 
 BUG FIXES:
 
-- `-PRI` suffix was added to the name of primary connection even if there was no secondary connection
+- `-PRI` suffix was added to the name of primary connection even if there was no secondary connection.
 
 ## 0.1.0 (March 25, 2022)